Convert NeverNull<T> to struct.
The struct will probably do a better job at fending off null,
even if it's possible to invoke default constructor on it and
thus making the encapsulated value null. Hopefull these
will be very rare occasions. And I put a guard in the Value
property to throw with a nice message.